requirements.txt
asgiref==3.2.3
Django==3.0.3
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
Markdown==3.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
